module modified

+++ app/Http/Controllers/team/team/ManageTeamController.php
@@ -15,6 +15,7 @@ use App\Services\TeamService;
 use Exception;
 use Image;
 use App\Models\Event\Event;
+use App\Models\Event\EventRegisterTeam;
 
 class ManageTeamController extends Controller
 {
@@ -241,10 +242,19 @@ class ManageTeamController extends Controller
 
     public function deleteteam($id) 
     {
-       $result= Team::find($id);
-       $result->delete();
-       parent::successMessage("Team deleted successfully");
-       return redirect()->back();
+        $results = EventRegisterTeam::where('team_id', $id)->first();
+        if(empty($results)) {
+            $result= Team::find($id);
+            $result->delete();
+            parent::successMessage("Team deleted successfully");
+            return redirect()->back();            
+        } else {
+            parent::dangerMessage("Team added into event");
+            parent::dangerMessage("You cannot delete this team");
+            return redirect()->back();         
+        }
+       
+      
 
     }
